Memorization in styla
This will save a lot of time when there are multiple ways to prove a
sub goal. Otherwise, the time is exponential to the inference depth.

When the parameters in goal head are all constants, or if all the parameters
in the goal doesn't appear in the tail, we can cache the result. As they
are facts, once styla proves it right or wrong, the result will never
change, and it doesn't affect the unification of any other variables.
When unfolding such goal head, add it to unfolded map that track in
which unfolder it was first unfolded. Keep the unfolded goal in the
goal list as a separator to determine when all its children are evaluated.
If we meet the "seperator" head again, that means it evaluates to true. Otherwise,
if all the clauses in the current unfolder has been used, and we still haven't
see successful result in cachedResult map, that means it evaluates to false.

Use objectHashCode for Var when determining if a term has been unfolded.
hashCode doesn't work. I saw collision when the predicate name is the same
and parameters are in different order.

Use hashCode when removing duplicate rules.